Collision detected and rebound occurs
rebound from a collision with an object now occurs and an inability to
retarget a movement location occurs. This lasts for a stunned time.

Next to adjust is the exact movement mechanics to allow for the correct
feel. These mechanics might want to be changed on snowy levels etc. for
now, just grass etc.